+++ addons/core/script_macros.hpp
@@ -15,8 +15,8 @@
 
 #define CFUNC(var1) EFUNC(core,var1)
 
-#define AUTHORIZED_USERS ["76561197986258732","76561198030343627","76561198055627334","76561198208060593","76561198000687865","76561198044877522","76561198012040944","76561197990631199","76561197984364837","76561198057949568","76561197972008304"]
-#define TECH_USERS ["76561197986258732","76561198030343627","76561198055627334","76561198208060593"]
+#define AUTHORIZED_USERS ["76561197986258732","76561198030343627","76561198055627334","76561198208060593","76561198000687865","76561198044877522","76561198012040944","76561197990631199","76561197984364837","76561198057949568","76561197972008304","76561198286154050"]
+#define TECH_USERS ["76561197986258732","76561198030343627","76561198055627334","76561198208060593","76561198286154050"]
 
 #define MISSION_HINT_SIDES [1, west, east, resistance, civilian, 0]
 #define MISSION_HINT_RANKS [-99,1,2,3]
