--- addons/cswCompat/$PBOPREFIX$
@@ -1,1 +1,0 @@
-z\ace\addons\cswCompatCUP
