--- addons/cswCompat/CfgMagazineGroups.hpp
@@ -1,34 +1,0 @@
-class ace_csw_groups {
-    class CUP_AT13_M {
-        CUP_6Rnd_AT13_M = 1;
-    };
-    class GVAR(cup_spg9_at) {
-        CUP_16Rnd_PG9_AT_M = 1;
-    };
-    class GVAR(cup_spg9_he) {
-        CUP_16Rnd_OG9_HE_M = 1;
-    };
-    class GVAR(cup_tow_at) {
-        CUP_6Rnd_TOW_HMMWV_M = 1;
-    };
-    class GVAR(20Rnd_30mm_G_belt) {
-        CUP_29Rnd_30mm_AGS30_M = 1;
-    };
-
-    // reloadable m119
-    class GVAR(105mm_m119_he) {
-        CUP_30Rnd_105mmHE_M119_M = 1;
-    };
-    class GVAR(105mm_m119_smoke) {
-        CUP_30Rnd_105mmSMOKE_M119_M = 1;
-    };
-    class GVAR(105mm_m119_wp) {
-        CUP_30Rnd_105mmWP_M119_M = 1;
-    };
-    class GVAR(105mm_m119_laser) {
-        CUP_30Rnd_105mmLASER_M119_M = 1;
-    };
-    class GVAR(105mm_m119_illum) {
-        CUP_30Rnd_105mmILLUM_M119_M = 1;
-    };
-};
