--- addons/cswCompat/script_component.hpp
@@ -1,14 +1,0 @@
-#define COMPONENT cswCompatCUP
-#include "\z\ace\addons\main\script_mod.hpp"
-
-// #define DEBUG_MODE_FULL
-// #define DISABLE_COMPILE_CACHE
-// #define ENABLE_PERFORMANCE_COUNTERS
-
-#include "\z\ace\addons\main\script_macros.hpp"
-
-// because this is a "ace" addon we have to include this manually
-#include "\z\potato\addons\core\script_macros_debug.hpp"
-
-#define ACE_CSW_PATH(XX) QUOTE(\z\ace\addons\csw\XX)
-#define ACE_APL_PATH(XX) QUOTE(\z\ace\addons\apl\XX)
