+++ addons/miscFixes/CfgVehicles.hpp
@@ -58,71 +58,6 @@ class CfgVehicles {
     // diag_log text format ["class %1: %2 { SLX_XEH_DISABLED = 1; }; // %3 -> %4", _vehType, _parent, _addon, getText (configFile >> "CfgVehicles" >> _vehType >> "displayName")];
     // } forEach x3;
 
-
-    // Fix: WARNING: MBG_WaterfallPlugin does not support Extended Event Handlers! Addon: @BourbonMapRotation
-    class Logic;
-    class MBG_WaterfallPlugin: Logic {
-        class EventHandlers {
-            class CBA_Extended_EventHandlers: CBA_Extended_EventHandlers {};
-        };
-    };
-
-    // Fix: Drongo's Spooks and Anomalies XEH
-    class DSA_SpookBase;
-    class DSA_SpookBase2;
-    class DSA_AnomalyBase;
-    class B_Soldier_F;
-    class DSA_Vampire: DSA_SpookBase2 { class EventHandlers { class CBA_Extended_EventHandlers: CBA_Extended_EventHandlers {}; }; }; // Drongos Spooks and Anomalies> Vampire
-    class DSA_Wendigo: DSA_SpookBase { class EventHandlers { class CBA_Extended_EventHandlers: CBA_Extended_EventHandlers {}; }; }; // Drongos Spooks and Anomalies> Wendigo
-    class DSA_Shadowman: DSA_SpookBase { class EventHandlers { class CBA_Extended_EventHandlers: CBA_Extended_EventHandlers {}; }; }; // Drongos Spooks and Anomalies> Shadowman
-    class DSA_Hatman: DSA_Shadowman { class EventHandlers { class CBA_Extended_EventHandlers: CBA_Extended_EventHandlers {}; }; }; // Drongos Spooks and Anomalies> Hatman
-    class DSA_Mindflayer: DSA_SpookBase2 { class EventHandlers { class CBA_Extended_EventHandlers: CBA_Extended_EventHandlers {}; }; }; // Drongos Spooks and Anomalies> Mindflayer
-    class DSA_411: DSA_SpookBase { class EventHandlers { class CBA_Extended_EventHandlers: CBA_Extended_EventHandlers {}; }; }; // Drongos Spooks and Anomalies> 411
-    class DSA_Rake: DSA_SpookBase { class EventHandlers { class CBA_Extended_EventHandlers: CBA_Extended_EventHandlers {}; }; }; // Drongos Spooks and Anomalies> Rake
-    class DSA_Abomination: DSA_411 { class EventHandlers { class CBA_Extended_EventHandlers: CBA_Extended_EventHandlers {}; }; }; // Drongos Spooks and Anomalies> Abomination
-    class DSA_Snatcher: DSA_SpookBase { class EventHandlers { class CBA_Extended_EventHandlers: CBA_Extended_EventHandlers {}; }; }; // Drongos Spooks and Anomalies> Snatcher
-    class DSA_Crazy: DSA_SpookBase2 { class EventHandlers { class CBA_Extended_EventHandlers: CBA_Extended_EventHandlers {}; }; }; // Drongos Spooks and Anomalies> Crazy
-    class DSA_Launchpad: DSA_AnomalyBase { class EventHandlers { class CBA_Extended_EventHandlers: CBA_Extended_EventHandlers {}; }; }; // Drongos Spooks and Anomalies> Launchpad
-    class DSA_Leech: DSA_AnomalyBase { class EventHandlers { class CBA_Extended_EventHandlers: CBA_Extended_EventHandlers {}; }; }; // Drongos Spooks and Anomalies> Leech
-    class DSA_Zapper: DSA_AnomalyBase { class EventHandlers { class CBA_Extended_EventHandlers: CBA_Extended_EventHandlers {}; }; }; // Drongos Spooks and Anomalies> Zapper
-    class DSA_Trapdoor: DSA_AnomalyBase { class EventHandlers { class CBA_Extended_EventHandlers: CBA_Extended_EventHandlers {}; }; }; // Drongos Spooks and Anomalies> Trapdoor
-    class DSA_DeltaX_Operator: B_Soldier_F { class EventHandlers { class CBA_Extended_EventHandlers: CBA_Extended_EventHandlers {}; }; }; // Drongos Spooks and Anomalies> Delta X Operator
-    class DSA_DeltaX_CBRN: B_Soldier_F { class EventHandlers { class CBA_Extended_EventHandlers: CBA_Extended_EventHandlers {}; }; }; // Drongos Spooks and Anomalies> Delta X CBRN
-
-    // Fix: FFAA
-    class FFAA_rapel_logica: Logic { SLX_XEH_DISABLED = 1; }; // @FFAA -> FFAA RAPEL
-    class Air;
-    class FFAA_rapel: Air { SLX_XEH_DISABLED = 1; }; // @FFAA -> Rappel rope
-    class StaticWeapon;
-    class FFAA_rapelf: StaticWeapon { SLX_XEH_DISABLED = 1; }; // @FFAA -> Rappel rope
-    class PlaneWreck;
-    class ffaa_ea_hercules_wreck: PlaneWreck { SLX_XEH_DISABLED = 1; }; // @FFAA -> Wreck
-    class ffaa_AV8BWreck: PlaneWreck { SLX_XEH_DISABLED = 1; }; // @FFAA -> Wreck
-    class ffaa_mq9_reaper_Wreck: PlaneWreck { SLX_XEH_DISABLED = 1; }; // @FFAA -> Wreck
-    class HelicopterWreck;
-    class ffaa_et_ch47Wreck: HelicopterWreck { SLX_XEH_DISABLED = 1; }; // @FFAA -> Wreck
-    class ffaa_famet_cougarWreck: HelicopterWreck { SLX_XEH_DISABLED = 1; }; // @FFAA -> Wreck
-
-    // Fix fajita
-	class Building;
-	class NonStrategic: Building {
-		class DestructionEffects;
-	};	
-	class HouseBase: NonStrategic {};	
-	class House: HouseBase {
-        class DestructionEffects: DestructionEffects {
-			class Smoke1 {
-				type="HouseDestructionSmoke3";
-			};
-			class Smoke2: Smoke1 {
-				type="HouseDestructionSmoke4";
-			};
-			class Smoke3: Smoke1 {
-				type="HouseDestrSmokeLong";
-            };
-		};
-	};
-
     // Fix AEGIS
     class C_man_1;
     class C_man_priest_F: C_man_1 { class EventHandlers { class CBA_Extended_EventHandlers: CBA_Extended_EventHandlers {}; }; };
