+++ addons/settings/$PBOPREFIX$
@@ -1,0 +1,1 @@
+z\potato\addons\settings
