+++ addons/settings/script_component.hpp
@@ -1,0 +1,12 @@
+#define COMPONENT settings
+#include "\z\potato\addons\core\script_mod.hpp"
+
+// #define DEBUG_MODE_FULL
+// #define DISABLE_COMPILE_CACHE
+// #define ENABLE_PERFORMANCE_COUNTERS
+
+#ifdef DEBUG_ENABLED_SETTINGS
+    #define DEBUG_MODE_FULL
+#endif
+
+#include "\z\potato\addons\core\script_macros.hpp"
