+++ addons/spectate/functions/fnc_toggleUI.sqf
@@ -22,6 +22,7 @@ GVAR(uiVisible) = !GVAR(uiVisible);
 if (GVAR(uiVisible)) then {
     FOCUS_GROUP ctrlSetFade 0;
     CAM_GROUP ctrlSetFade 0;
+	GCAM ctrlSetFade 0;
     LIST ctrlSetFade 0.8;
     [GVAR(respawnOpen)] call FUNC(setRespawn);
 
@@ -30,6 +31,7 @@ if (GVAR(uiVisible)) then {
 } else {
     FOCUS_GROUP ctrlSetFade 1;
     CAM_GROUP ctrlSetFade 1;
+	GCAM ctrlSetFade 1;
     LIST ctrlSetFade 1;
     [false] call FUNC(setRespawn);
 
@@ -39,4 +41,5 @@ if (GVAR(uiVisible)) then {
 
 FOCUS_GROUP ctrlCommit 0.25;
 CAM_GROUP ctrlCommit 0.25;
+GCAM ctrlCommit 0.25;
 LIST ctrlCommit 0.25;
